requirements.txt
Django>=3.2,<4.0
gunicorn==20.1.0
python-dotenv